fix: byte encoder for kafka messages
refactored common logic into new codec class
