+++ mlserver/kafka/message.py
@@ -1,28 +1,7 @@
-import json
-
-from typing import Dict, Optional, List, Tuple, Union
+from typing import Dict, Optional, List, Tuple
 
 from pydantic import BaseModel
-
-try:
-    import orjson
-except ImportError:
-    orjson = None  # type: ignore
-
-
-def _encode_value(v: dict) -> bytes:
-    if orjson is None:
-        dumped = json.dumps(v)
-        return dumped.encode("utf-8")
-
-    return orjson.dumps(v)
-
-
-def _decode_value(v: Union[bytes, str]) -> dict:
-    if orjson is None:
-        return json.loads(v)
-
-    return orjson.loads(v)
+from ..codecs.json import encode_to_json_bytes, decode_from_bytelike_json_to_dict
 
 
 def _encode_headers(h: Dict[str, str]) -> List[Tuple[str, bytes]]:
@@ -48,7 +27,7 @@ class KafkaMessage(BaseModel):
     @classmethod
     def from_kafka_record(cls, kafka_record) -> "KafkaMessage":
         key = kafka_record.key
-        value = _decode_value(kafka_record.value)
+        value = decode_from_bytelike_json_to_dict(kafka_record.value)
         headers = _decode_headers(kafka_record.headers)
         return KafkaMessage(key=key, value=value, headers=headers)
 
@@ -61,7 +40,7 @@ class KafkaMessage(BaseModel):
 
     @property
     def encoded_value(self) -> bytes:
-        return _encode_value(self.value)
+        return encode_to_json_bytes(self.value)
 
     @property
     def encoded_headers(self) -> List[Tuple[str, bytes]]:
